graph: split the graph in to a seperate package from the builder
